+++ omniqueue/src/lib.rs
@@ -101,8 +101,8 @@ mod scheduled;
 #[allow(deprecated)]
 pub use self::{
     queue::{
-        Acker, BaseDynConsumer, BaseDynProducer, Delivery, DynConsumer, DynProducer, QueueConsumer,
-        QueueProducer,
+        Acker, BaseDynConsumer, BaseDynProducer, BatchResult, Delivery, DynConsumer, DynProducer,
+        QueueConsumer, QueueProducer,
     },
     scheduled::{BaseDynScheduledProducer, DynScheduledProducer, ScheduledQueueProducer},
 };
@@ -139,8 +139,8 @@ pub enum QueueError {
 }
 
 impl QueueError {
-    pub fn generic<E: std::error::Error + Send + Sync + 'static>(e: E) -> Self {
-        Self::Generic(Box::new(e))
+    pub fn generic(e: impl Into<Box<dyn std::error::Error + Send + Sync>>) -> Self {
+        Self::Generic(e.into())
     }
 }