+++ CHANGELOG.md
@@ -6,6 +6,11 @@
 - Delete the `QueueBuilder` type and `QueueBackend` trait
   - The in-memory backend is now constructed using `InMemoryBackend::new_pair()`
   - Every other backend now has its own builder type now
+- Change signature of `Error::generic` to allow more parameter types (use
+  `impl Into<Box<dyn std::error::Error + Send + Sync>>` instead of
+  `impl std::error::Error + Send + Sync + 'static`)
+- The batch-sending methods now return `Result<BatchResult>` instead of `Result<()>`,
+  to be able to represent partially-successful batch send operations
 
 # 0.3.0